% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bas.point.r
\name{bas.point}
\alias{bas.point}
\title{Draws a Balanced Asseptance Sample (BAS) from a discrete resource (points).}
\usage{
bas.point(x, n, ...)
}
\arguments{
\item{x}{A SpatialPoints or SpatialPointsDataFrame object. This object
must contain at least 1 point.}

\item{n}{Sample size.  Number of points to select from the set of points
contained in \code{x}.}
}
\value{
A SpatialPointsDataFrame containing locations in the BAS sample, in
order they are to be visited.  A 'siteID' attribute is attached to each
point (in the embedded data frame) and gives the BAS ordering of the sample
(i.e., sort on 'siteID' to get proper BAS order).  In addition, if the input
object has an attached data frame (i.e., is a SpatialPointsDataFrame), the
attrributes of the point are attached in the associated data frame.
}
\description{
Draws a BAS sample from a SpatialPoints object
}
\details{
The BAS method for points computs the minimum distance between all points in
\code{x} and places a small square (pixel) around each.  Size of the
square around each point is d/sqrt(2) on a side, where d is the minimum
distance between points. The BAS method for points then selects a BAS sample
from the set of polygons (i.e., squares) surrounding each point (see
\code{\link{bas.polygon}}).  The BAS method of polygons selects Halton
points until \code{n} points are located inside the squares surrounding the
points.  When a square contains a Halton point, the official sample location
is the the original point (center of the square), not the Halton point.
}
\examples{

\dontrun{
bas.point( WA.cities, 100)
}



}
\author{
Trent McDonald
}
\seealso{
\code{\link{bas.polygon}}, \code{\link{bas.line}}, \code{\link{spsample}}
}
\keyword{design}
\keyword{survey}

