\name{cleanup}
\alias{cleanup}

\title{
SDraw clean up}
\description{
Clean up the user workspace.}
\usage{
cleanup(x)
}

\details{
Closes the SDraw window. Erases all hidden objects placed in .GlobalEnv by SDraw. 
}
\value{
NULL.  This is run for its side effects.}
\author{
Trent McDonald}
\examples{
cleanup()}
\keyword{ modeling }

