% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.line.r
\name{hal.line}
\alias{hal.line}
\title{Draws a Halton Lattice sample from a linear (line) resource .}
\usage{
hal.line(x, n, J = NULL, pt.spacing = NULL, bases = c(2, 3),
  plot = TRUE, plot.lattice = FALSE)
}
\arguments{
\item{x}{A SpatialLines or SpatialLinesDataFrame object. This object must
contain at least 1 line.  If it contains more than 1 line, the HAL sample is
drawn from the union of all lines.}

\item{n}{Sample size.  Number of locations to draw from the set of all lines
contained in \code{x}.}

\item{J}{A 2X1 vector of base powers.  \code{J[1]} is for horizontal,
\code{J[2]} for vertical dimension. \code{J} determines the size and shape
of the smallest Halton boxes. There are \code{bases[1]^J[1]} vertical columns 
of Halton boxes over \code{x}'s bounding box, and \code{bases[2]^J[2]} 
horizontal rows of Halton boxes over the bounding box, for a total 
of \code{prod(bases^J)} total boxes.  The dimension of each box is 
\code{c(dx,dy)/(bases^J)}, where \code{c(dx,dy)} are the horizontal and 
vertical extents of \code{x}'s bounding box.  If \code{J=NULL} (the default),
\code{J} is choosen so that Halton boxes are as square as possible.}

\item{pt.spacing}{Desired spacing of points on lines prior to sampling via
HAL.  The first step in sampling is descritization which places
equaly-spaced points on all lines. Then, points are sampled using Halton sampling
(see \code{hal.point}) for points.  This parameter controls spacing of points during the
discretization of lines.  For example, specifying 50, and assuming
\code{x} is projected to UTM meters, means points will be placed every 50
meters along all lines in \code{x}. \code{x} should be projected before
sampling so that \code{pt.spacing} makes sense.  If \code{pt.spacing} is not
specified, 1000*\code{n} points will be placed along the lines during
descretization.}

\item{bases}{2X1 vector of Halton bases.  These must be co-prime.}

\item{plot}{Logical indicating whether to plot \code{x} and selected points.}

\item{plot.lattice}{Logical indicating whether to plot the Halton lattice used 
to draw the sample.  \code{plot.lattice = TRUE} produces same map 
as \code{plot=TRUE}, with the addition of the Halton lattice.}
}
\value{
A \code{SpatialPointsDataFrame} containing locations in the HAL sample, in
the order they are to be visited.  A 'siteID' attribute is attached to each
point (in the embedded data frame) and gives the HAL ordering of the sample
(i.e., sort on 'siteID' to get proper HAL order).  In addition, if the input
object has an attached data frame (i.e., is a  \code{SpatialLinessDataFrame}), the
attrributes of the line on which each HAL point fell is attached in the
associated data frame. The number of the line in \code{x} on which each
point falls is an attribute of the output points.
}
\description{
Draws a Halton Lattice sample from a SpatialLines object, where SpatialLines
are defined in library 'sp'.
}
\details{
A HAL sample is drawn from the union of all lines in \code{x} by first
discretizing lines with points spaced \code{pt.spacing} apart. After
discretizing the lines, discretization points are sampled using the HAL method
for points (see \code{\link{hal.point}}).
}
\examples{

  #   Default sample of Hawaii coastline
  samp <- hal.line( 100, HI.coast )
  
  #   Frame spacing ~157 meters. Frame has ~10,000 points
  samp <- hal.line( 100, HI.coast, pt.spacing=157.1855, plot.lattice=TRUE )
  
  #   Different Halton lattice 
  samp <- hal.line( 100, HI.coast, J=c(5,3), pt.spacing=157.1855, plot.lattice=TRUE)

}
\author{
Trent McDonald
}
\seealso{
\code{\link{bas.line}}, \code{\link{hal.point}}, \code{\link{hal.polygon}}, \code{\link{spsample}}
}
\keyword{design}
\keyword{survey}

