% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.polygon.r
\name{hal.polygon}
\alias{hal.polygon}
\title{Draws a Halton Lattice sample from an area (polygon) resource.}
\usage{
hal.polygon(n, shp, J = NULL, eta = c(1, 1), triangular = FALSE,
  bases = c(2, 3))
}
\arguments{
\item{n}{Sample size.  Number of locations to draw from the set of points
contained in \code{shp}.}

\item{shp}{A SpatialPointss or SpatialPointssDataFrame object. This object must
contain at least 1 point.}

\item{J}{A 2X1 vector of base powers.  \code{J[1]} is for horizontal,
\code{J[2]} for vertical dimension. \code{J} determines the size and shape
of the smallest Halton boxes. There are \code{bases[1]^J[1]} vertical columns 
of Halton boxes over \code{shp}'s bounding box, and \code{bases[2]^J[2]} 
horizontal rows of Halton boxes over the bounding box, for a total 
of \code{prod(bases^J)} total boxes.  The dimension of each box is 
\code{c(dx,dy)/(bases^J)}, where \code{c(dx,dy)} are the horizontal and 
vertical extents of \code{shp}'s bounding box.  If \code{J=NULL} (the default),
\code{J} is choosen so that Halton boxes are as square as possible.}

\item{eta}{A 2X1 vector specifying the number of points to add in the 
horizontal and vertical dimensions of each Halton box.  e.g., if 
\code{eta} = c(3,2), a grid of 3 (horizontal) by 2 (vertical) points is 
added inside each Halton box.}

\item{triangular}{A boolean scalar. If TRUE, odd horizontal rows of the 
Halton lattice are moved one-quarter a Halton box width to the right, while 
even rows of the Halton lattic are moved one-quarter of a Halton box width 
to the left. This creates a triangular Halton lattice over the 
bounding box of the polygons. 
If FALSE, a rectangluar Halton lattice is constructed.}

\item{bases}{2X1 vector of Halton bases.  These must be co-prime.}
}
\value{
A \code{SpatialPointsDataFrame} containing locations in the HAL sample, in
the order they are to be visited.  A 'siteID' attribute is attached to each
point (in the embedded data frame) and gives the HAL ordering of the sample
(i.e., sort on 'siteID' to get proper HAL order).  In addition, if the input
object has an attached data frame (i.e., is a \code{SpatialPolygonsDataFrame}), the
attributes of the polygon in which each HAL point fell is attached in the
associated data frame. The ID  of the polygon in \code{shp} in which each
point fell is an attribute of the output points.
}
\description{
Draws a Halton Lattice sample from a \code{SpatialPolygons} object.
}
\details{
\emph{Brief description of Halton Lattice sampling for polygons:} 
Given a set of Halton Lattice parameters \code{J}, \code{bases}, \code{eta}, 
and  \code{triangular}, the union of all polygons is discretized by placing 
a Halton lattice of \code{prod(bases^J)} boxes, each containing  \code{prod(eta)}
points, over the bounding box of all polygons. Points outside any polygon are
dropped, and the resulting Halton lattice frame is passed to \code{hal.point} 
for sampling.
}
\note{
The number of points in the Halton lattice becomes large very quickly. 
The number of points in the Halton lattice is \code{prod(bases^J)*prod(eta)}.
}
\examples{

#   Draw sample of Hawaii coastline
#   This takes approximately 30 seconds to run
samp <- hal.polygon( 100, WA )
plot(WA)
points( samp, pch=16, col="red" )
  
#   Different lattice topology
samp <- hal.polygon( 100, WA, J=c(8,2), eta=c(2,1), triangular=TRUE)
points( samp, pch=15, col="blue")
  

}
\author{
Trent McDonald
}
\seealso{
\code{\link{hal.line}}, \code{\link{hal.point}}, \code{\link{spsample}}
}
\keyword{design}
\keyword{survey}

