% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hip.lattice.point.r
\name{hip.lattice.point}
\alias{hip.lattice.point}
\title{Halton Iterative Partition lattice inside a \code{bbox} (bounding box) matrix object.}
\usage{
hip.lattice.point(coords, J, bases, ...)
}
\arguments{
\item{coords}{A \code{data.frame} of the coordinates from the sample space.}

\item{J}{A 2X1 vector of base powers which determines the size and shape 
of the Halton boxes. See additional description in help for 
\code{\link{hip.polygon}} function.}

\item{bases}{A 2X1 vector of Halton bases.  These must be co-prime.}
}
\value{
A \code{list} of \code{data.frames} containing locations in the Halton lattice of the 
partitioned boxes
}
\description{
Constructs an iteratively partitioned lattice of Halton boxes (a Halton lattice) inside a 
bounding box \code{bbox} of the sample space. This method does the 
hard work of partitioning the boxes to sample from. It is meant to be used internally by 
\code{hip.point} only.
}
\details{
This routine is called internally by \code{hip.point}, and is not 
normally called by the user. This should be avoided.
}
\examples{

# Take a simple HIP lattice for illustration
# make a plot for test
plot(x = c(0,1), y = c(0,1), xlim = c(0,1), ylim=c(0,1))

# create data.frame of points to plot
pnts <- matrix(c(0.2, 0.2,
                 0.1, 0.5,
                 0.3, 0.9,
                 1.0, 0.5,
                 0.9, 0.6,
                 0.7, 0.7), nrow = 6, byrow = T)
pnts <- data.frame(pnts)

# plot data.frame of points
points(pnts)

# Test of hip.lattice.point.r
hip.lattice.point(coords = data.frame(pnts), J = c(1,1), bases = c(2,3)) 
# output ordering of points are Halton order

}
\seealso{
\code{\link{hip.point}}, \code{\link{hip.polygon}}, \code{\link{hip.lattice.polygon}}
}
\author{
Michael J Kleinsasser
}
\keyword{design}
\keyword{survey}
