% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hip.plot.lattice.r
\name{hip.plot.lattice}
\alias{hip.plot.lattice}
\title{Plot a Halton Lattice over a polygon resource}
\usage{
hip.plot.lattice(resource, bases = c(2, 3), J = c(8, 5),
  sample = NULL)
}
\description{
Plot a Halton Lattice over a polygon resource
}
