% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hip.point.r
\name{hip.point}
\alias{hip.point}
\title{Draws a Halton Iterative Partition (HIP) sample from a discrete 
2-dimensional (point) resource.}
\usage{
hip.point(x, n)
}
\arguments{
\item{x}{A \code{SpatialPoints} or \code{SpatialPointsDataFrame} object. 
This object must contain at least 1 point. \code{x} is the 2-dimensional point resource
from which samples are taken.}

\item{n}{Target sample size.  Target number of locations to draw from the set of points
contained in \code{x}. If the sample size returned is less than the desired sample size,
increase \code{n} until the desired sample size is reached.}
}
\value{
A \code{SpatialPoints} containing locations in the HIP sample.

Additional attributes of the output object, beyond those which 
make it a \code{SpatialPoints}, are:
\itemize{
   \item \code{frame}: Name of the input sampling frame.
   \item \code{frame.type}: Type of resource in sampling frame. (i.e., "point").
   \item \code{sample.type}: Type of sample drawn. (i.e., "HIP").
   
   \item \code{J}: Exponents of the bases used to form the lattice of 
   Halton boxes. This is either the input \code{J}, or the \code{J} vector
   computed by \code{\link{halton.indices}}.  
   
   \item \code{bases}: Bases of the Halton sequence used to draw the sample. 
   
   \item \code{hl.bbox}: The bounding box around points in \code{x} used to 
   draw the sample.  See \code{\link{halton.indices}}.
 
}
}
\description{
Draws a Halton Iterative Partition (HIP) sample from a 
\code{SpatialPoints*} object.
}
\details{
\bold{A brief description of Halton Iterative Partition (HIP) sampling for points:} 
Given a set of Halton Iterative Partition parameters \code{x} (SpatialPoints* object) and \code{n} (sample size),
a lattice of Halton boxes is constructed iteratively over the bounding box of the input points.  
This results in enough Halton boxes on the bounding box to uniquely cover all points in the point resource (one box
for each point in the point resource). 
The Halton index of all boxes is computed and assigned to points that lie 
in each box. Finally, a random number between 0 and the largest Halton index is 
drawn, and the next \code{n} coordinates in the mapped real numbers are taken as 
the sample.
}
\examples{

#   Draw sample of cities in the state of Washington
data(WA.cities)
samp <- hip.point( WA.cities, 100 )
  

}
\seealso{
\code{\link{hip.line}}, \code{\link{hip.polygon}}, \code{\link{SDraw}}, 
\code{\link{bas.point}}
}
\author{
Michael J. Kleinsasser
}
\keyword{design}
\keyword{survey}
