% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.U.r
\name{maxU}
\alias{maxU}
\title{Maximum integer used in the BAS routines}
\usage{
maxU()
}
\value{
10e7 or 100,000,000
}
\description{
A function that returns the maximum integer used 
to construct the random-start Halton sequences.  By redefining 
this function and placing in the .GlobalEnv environment, the
user can change the maximum integer and hence the number of 
possible BAS samples.
}
\examples{
# A 2D random-start Halton sequence, length 10, bases c(2,3).
u <- c( floor((maxU()+1)*runif(1)), floor((maxU()+1)*runif(1)))
halt.pts <- halton(10,dim=2,start=u,bases=c(2,3))


}
\author{
Trent McDonald
}
\references{
Robertson, B.L., J. A. Brown, T. L. McDonald, and P. Jaksons
(2013) "BAS: Balanced Acceptance Sampling of Natural Resources", Biometrics,
v69, p. 776-784.
}
\seealso{
\code{\link{bas.line}}, \code{\link{bas.point}}, \code{\link{bas.polygon}}
}

