% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.samp.R
\name{plot.samp}
\alias{plot.samp}
\title{Plot sample and frame}
\usage{
plot.samp(x, frame, lattice = FALSE, bbox = FALSE, background = NULL)
}
\arguments{
\item{x}{A \code{SpatialPointsDataFrame} produced by an SDraw sampling function.  For example, 
as produced by \code{sdraw(frame,n)}.  This object is a standard \code{SpatialPointsDataFrame} object
with additional attributes that record the sampling design.}

\item{frame}{The sample frame used to draw the sample contained in \code{x}. This is 
either a \code{SpatialPoints*}, \code{SpatialLines*}, or \code{SpatialPolygons*} object.}

\item{lattice}{Boolean. Whether to plot the Halton Lattice if \code{x} is a HAL sample.}

\item{bbox}{Boolean. Whether to plot the bounding box if the sample 
has a bounding box attribute. This generally means \code{x} 
is a HAL or BAS sample.  \code{lattice==TRUE} sets \code{bbox == TRUE}.}

\item{background}{Either the background image (from a previous call to this routine) or 
the type of background image to plot.  If NULL (the default), no background image is plotted. 
If \code{background} is a string, it names the type of background image 
to retrieve from the Open Maps project.  Background types are 
any allowed by \code{\link{openmap}}  in the OpenStreetMap package.  Popular options are "osm", 
"esri", and "esri-topo".  When a background is retrieved, it slows execution and requires 
an internet connection.  When \code{background} is the returned value from a previous call 
to this function, it is plotted without retrieving a new image.  This will speed multiple 
iterations of plotting when the sample frame does not change.}
}
\value{
The set of tiles used in the background image is silently returned. If 
\code{background==NULL}, NULL is silently returned.
}
\description{
Plot the sample and optionally the frame, background image (terrain), and lattice 
 (if HAL sample).
}
\examples{
data(WY)
samp <- sdraw(WY, 100, type="HAL", J=c(4,3))
plot.samp( samp, WY )


}
\author{
Trent McDonald
}
\seealso{
\code{\link{sdraw}}
}

