% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLattice.R
\name{plotLattice}
\alias{plotLattice}
\title{plotLattice}
\usage{
plotLattice(latt, indices = NULL, J = NULL, bases = c(2, 3),
  box = NULL, ...)
}
\arguments{
\item{latt}{A list containing the Halton boxes in the lattice. 
One item in the list for each Halton box.  Each list item is a 
2 X 2 matrix where first row is min and max of dimension 1, 
second row is min and max of dimension 2.}

\item{indices}{The halton indices of all boxes.  \code{length(latt)} must equal
\code{length(indices)} and \code{indices[i]} is the index of box \code{latt[[i]]}. 
If missing or NULL, indices are not printed.}

\item{box}{The bounding box of all the halton boxes in \code{latt}. If missing
or NULL, the min and max extent of \code{latt} are used.}

\item{...}{Options of \code{graphics::lines} that control appearance of 
the boxes.  For example, 
\code{lty}, \code{col}, etc.}
}
\value{
NULL is returned invisibly.
}
\description{
Plot a Halton lattice
}
\details{
Lines on the current plot are produced.
}
\author{
Trent McDonald

  @examples 
latt <- hip.lattice.polygon( box = matrix(c(0,0,1,1),2), J = c(2,2) bases = c(2,3) )
plotLattice(latt)
}
