% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUI.r
\name{runUI}
\alias{runUI}
\title{Launch SDraw Shiny interface}
\usage{
runUI()
}
\value{
None.
}
\description{
Launches the Shiny interface to SDraw which allows samples to 
 be drawn using a graphical user interface (UI)
}
\details{
Used for side effects.  A HTML viewer is launched. Control and 
use of the UI should be self-explanatory.
}
\examples{
\dontrun{runUI()}

}

