% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.point.r
\name{hal.point}
\alias{hal.point}
\title{Draws a Halton Lattice sample from a discrete (point) resource.}
\usage{
hal.point(x, n, J = NULL, bases = c(2, 3), plot = TRUE,
  plot.lattice = FALSE)
}
\arguments{
\item{n}{Sample size.  Number of locations to draw from the set of points
contained in \code{shp}.}

\item{J}{A 2X1 vector of base powers.  \code{J[1]} is for horizontal,
\code{J[2]} for vertical dimension. \code{J} determines the size and shape
of the smallest Halton boxes. There are \code{bases[1]^J[1]} vertical columns 
of Halton boxes over \code{shp}'s bounding box, and \code{bases[2]^J[2]} 
horizontal rows of Halton boxes over the bounding box, for a total 
of \code{prod(bases^J)} total boxes.  The dimension of each box is 
\code{c(dx,dy)/(bases^J)}, where \code{c(dx,dy)} are the horizontal and 
vertical extents of \code{shp}'s bounding box.  If \code{J=NULL} (the default),
\code{J} is choosen so that Halton boxes are as square as possible.}

\item{bases}{2X1 vector of Halton bases.  These must be co-prime.}

\item{plot}{Logical indicating whether to plot \code{shp} and selected points.}

\item{plot.lattice}{Logical indicating whether to plot the Halton lattice used 
to draw the sample.  \code{plot.lattice = TRUE} produces same map 
as \code{plot=TRUE}, with the addition of the Halton lattice.}

\item{shp}{A SpatialPointss or SpatialPointssDataFrame object. This object must
contain at least 1 point.}
}
\value{
A SpatialPointsDataFrame containing locations in the HAL sample, in
the order they are to be visited.  A 'SDraw.siteID' attribute is attached to each
point (in the embedded data frame) and gives the HAL ordering of the sample
(i.e., sort on 'siteID' to get proper HAL order).  In addition, if the input
object has an attached data frame (i.e., is a  \code{SpatialPointsDataFrame}), the
attrributes of the line on which each HAL point fell is attached in the
associated data frame. The number of the line in \code{shp} on which each
point falls also appears in the attribute data frame.
}
\description{
Draws a Halton Lattice sample from a \code{SpatialPoints*} object.
}
\details{
Given a set of Halton Lattice parameters \code{J} and \code{bases},
a Halton lattice is constructed over the bounding box of the input points.  
This results in \code{prod(bases^J)} Halton boxes on the bounding box. 
The Halton index of all boxes is computed and assigned to points that lie 
in each box.  Points that lie in the same Halton box are randomly assigned 
unique Halton cycle numbers. This separates points in the same Halton box by
at least \code{prod(bases^J)} units when indicies are mapped to the real line. 
Finally, a random number between 1 and the largest Halton (index+cycle) is 
drawn, and the next \code{n} units in the mapped real numbers are taken as 
the sample, restarting from the beginning if necessary.
}
\examples{

#   Draw sample of Hawaii coastline
#   This takes approximately 30 seconds to run
samp <- hal.point( 100, WA.cities, plot.lattice=TRUE )

#   Different lattice topology
samp <- hal.point( 100, WA.cities, J=c(10,4), plot.lattice=TRUE)
  

}
\author{
Trent McDonald
}
\seealso{
\code{\link{hal.line}}, \code{\link{hal.polygon}}, \code{\link{spsample}}, \code{\link{bas.point}}
}
\keyword{design}
\keyword{survey}

