% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hip.plot.lattice.r
\name{hip.plot.lattice}
\alias{hip.plot.lattice}
\title{Plot a Halton Lattice over a polygon resource}
\usage{
hip.plot.lattice(resource, bases = c(2, 3), J = c(8, 5), sample = NULL)
}
\arguments{
\item{resource}{\code{SpatialPolygons} object. The resource the halton lattice will be plotted over.}

\item{bases}{The bases of the halton lattice. Should be equal to those used to draw the sample.}

\item{J}{The orders of the halton lattice. Should be equal to those used to draw the sample.}

\item{sample}{The HIP sample to be plotted with the lattice.}
}
\description{
Plots a halton lattice over a polygon resource. Primarily for demonstration of HIP sampling theory.
}
